+++ tests/pwsh/fixtures/pas/Unformatted.pas
@@ -1,0 +1,7 @@
+unit Unformatted;
+
+interface
+
+implementation
+
+end.